begin work on registration
